New tool option --ssl-gnore-revoke-offline. New value  CURLSSLOPT_NO_REVOKE for CURLOPT_SSL_OPTIONS.
The new option applies only to Schannel where we have automatic
certificate revocation checking by default. It is supposed to
ignore errors when direct access to CRL and OSCP endpoint is blocked
in environments with proxies. It can also anticipate the behaviour
of low-availability CRL servers. It is different from the
--ssl-no-revoke command which disables the revocation checks.